+++ src/space.h
@@ -49,14 +49,14 @@ struct fof_CoM {
 
 struct fof {
 
-  int *group_index;
-  int *group_size;
+  size_t *group_index;
+  size_t *group_size;
   double *group_mass;
   struct fof_CoM *group_CoM;
   int num_groups;
-  int min_group_size;
-  int group_id_default;
-  int group_id_offset;
+  size_t min_group_size;
+  size_t group_id_default;
+  size_t group_id_offset;
   int group_links_size_default;
   char base_name[PARSER_MAX_LINE_SIZE];
